Fix CI and UTs
